requirements.txt
pyyaml
google-generativeai
python-dotenv

